+++ backend/src/main/java/com/vphilip/finance/app/account/repository/AccountTypeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface AccountTypeRepository extends ListCrudRepository<AccountType, Integer> {
+public interface AccountTypeRepository extends ListCrudRepository<AccountType, Long> {
     @Modifying
     @Query("INSERT INTO account_type (id, label, description) " +
             "VALUES (:#{#accountType.id}, :#{#accountType.label}, :#{#accountType.description})")
